consolidate tests from test_req to test_req_file
